requirement.txt
supabase
razorpay
requests
python-dotenv
